+++ pokemon-app/src/components/SinglePokemon.js
@@ -7,8 +7,8 @@ import ThemeSongs from "./ThemeSongs";
 import homeSong from "../assets/homeSong.mp3";
 import Types from "./Types";
 import PokemonCry from "./PokemonCry";
+import MovesList from "./MovesList";
 import { useState } from "react";
-import { useScroll } from "../hooks";
 
 const SinglePokemon = ({
   pokemon,
@@ -18,45 +18,10 @@ const SinglePokemon = ({
   isPokeballRendering,
 }) => {
   const [statsOnTop, setStatsOnTop] = useState(false);
-  const [isScrollBottom, onScroll, scrollRef, scrollTop] = useScroll();
 
   // This checks if the moves property exist in pokemon object.
   const checkDoesMovesKeyExistInObject = () =>
     Object.prototype.hasOwnProperty.call(pokemon, "moves");
-  // This checks if the moves list length is less than or equal to 7.
-  const checkIsMovesListLengthSevenOrLess = () => pokemon?.moves.length <= 7;
-  // This checks if the stats list has been selected.
-  const checkIsStatsListSelected = () => statsOnTop;
-  // This checks if the scrollbar is at the top.
-  const checkIsScrollbarAtTop = () => scrollTop < 1;
-  // This checks if the scrollbar is at the bottom.
-  const checkIsScrollbarAtBottom = () => isScrollBottom;
-
-  /**
-   * To render the top-arrow, this function checks to see if the pokemon moves exist AND the scrollbar is not at the top AND the stats list has not been selected.
-   * @returns {boolean} A boolean.
-   */
-  const renderMovesListTopArrow = () =>
-    // True if the moves property exist in pokemon object.
-    checkDoesMovesKeyExistInObject() &&
-    // True if the scrollbar is not at the top.
-    !checkIsScrollbarAtTop() &&
-    // True if the stats list has not been selected.
-    !checkIsStatsListSelected();
-
-  /**
-   * To render the bottom-arrow, this function checks to see if the pokemon moves exist AND its moves length is not seven or less AND the scrollbar is not at the bottom AND the stats list has not been selected.
-   * @returns {boolean} A boolean.
-   */
-  const renderMovesListBottomArrow = () =>
-    // True if the moves property exist in pokemon object.
-    checkDoesMovesKeyExistInObject() &&
-    // True if the moves length is greater than seven.
-    !checkIsMovesListLengthSevenOrLess() &&
-    // True if the scrollbar is not at the bottom.
-    !checkIsScrollbarAtBottom() &&
-    // True if the stats list has not been selected.
-    !checkIsStatsListSelected();
 
   return (
     <>
@@ -118,36 +83,12 @@ const SinglePokemon = ({
                 Your Pokemon
               </div>
 
-              <div
-                className={`moves-box ${hasEnemy ? "has-enemy" : ""} ${
-                  statsOnTop ? "" : "on-top"
-                }`}
-                ref={scrollRef}
-                onScroll={onScroll}
-              >
-                <ol id="moves-list">
-                  {!pokemon?.moves
-                    ? null
-                    : pokemon?.moves.map((moveData, moveIndex) => (
-                        <li key={moveIndex}>{moveData.move.name}</li>
-                      ))}
-                </ol>
-              </div>
-
-              {renderMovesListTopArrow() && (
-                <span
-                  className={`moves-list--up-arrow ${
-                    hasEnemy ? "has-enemy" : ""
-                  }`}
-                ></span>
-              )}
-
-              {renderMovesListBottomArrow() && (
-                <span
-                  className={`moves-list--down-arrow ${
-                    hasEnemy ? "has-enemy" : ""
-                  }`}
-                ></span>
+              {checkDoesMovesKeyExistInObject() && (
+                <MovesList
+                  pokemonMoves={pokemon?.moves}
+                  hasEnemy={hasEnemy}
+                  statsOnTop={statsOnTop}
+                />
               )}
             </div>
             <div className={`pokemon-types ${hasEnemy ? "has-enemy" : ""}`}>
