+++ pokemon-app/src/styles/EnemyPokemon.css
@@ -49,7 +49,6 @@
   width: 100%;
   height: 95%;
 }
-.enemy-pokemon__moves-title,
 .enemy-pokemon__stats-title {
   position: absolute;
   font-size: 1.8em;
@@ -67,8 +66,7 @@
   z-index: 0;
 }
 
-.enemy-pokemon__stats,
-.enemy-pokemon__moves {
+.enemy-pokemon__stats {
   z-index: 1;
   box-sizing: border-box;
   font-size: 1em;
@@ -88,17 +86,14 @@
   background-color: rgb(160, 212, 212);
   z-index: 0;
 }
-.enemy-pokemon__stats::-webkit-scrollbar,
-.enemy-pokemon__moves::-webkit-scrollbar {
+.enemy-pokemon__stats::-webkit-scrollbar {
   width: 8.6px;
   height: 7.6px;
 }
-.enemy-pokemon__stats::-webkit-scrollbar-track,
-.enemy-pokemon__moves::-webkit-scrollbar-track {
+.enemy-pokemon__stats::-webkit-scrollbar-track {
   background: transparent;
 }
-.enemy-pokemon__stats::-webkit-scrollbar-thumb,
-.enemy-pokemon__moves::-webkit-scrollbar-thumb {
+.enemy-pokemon__stats::-webkit-scrollbar-thumb {
   background-color: rgba(0, 0, 0, 0.18);
   border-radius: 50px;
   border: 1px solid rgba(0, 0, 0, 0.5);
@@ -110,13 +105,11 @@
   z-index: 3;
 }
 
-.enemy-pokemon__stats-list,
-.enemy-pokemon__moves-list {
+.enemy-pokemon__stats-list {
   margin: 5px;
 }
 
-.enemy-pokemon__stats-list > li,
-.enemy-pokemon__moves-list > li {
+.enemy-pokemon__stats-list > li {
   font-size: 1em;
   margin: 10px;
 }
@@ -129,7 +122,6 @@
   transform: scale(1.4);
 }
 
-.enemy-pokemon__moves-title:hover,
 .enemy-pokemon__stats-title:hover {
   cursor: pointer;
 }
@@ -165,61 +157,4 @@
 
 .user-poke-none {
   display: none;
-}
-
-.enemy-moves-list--up-arrow {
-  position: absolute;
-  top: 11%;
-  left: 45.5%;
-  width: 0;
-  height: 0;
-  border-left: 7px solid transparent;
-  border-right: 7px solid transparent;
-  border-bottom: 7px solid #b11501;
-  z-index: 3;
-  animation-duration: 0.8s;
-  animation-iteration-count: infinite;
-  animation-timing-function: ease-in-out;
-  animation-name: uparrowmovesenemy;
-}
-
-.enemy-moves-list--down-arrow {
-  position: absolute;
-  top: 38%;
-  left: 45.5%;
-  width: 0;
-  height: 0;
-  border-left: 7px solid transparent;
-  border-right: 7px solid transparent;
-  border-top: 7px solid #b11501;
-  z-index: 3;
-  animation-duration: 0.8s;
-  animation-iteration-count: infinite;
-  animation-timing-function: ease-in-out;
-  animation-name: downarrowmovesenemy;
-}
-
-/**************/
-/* ANIMATIONS */
-/**************/
-
-/* ENEMY-MOVES-LIST ARROWS MOVES */
-@keyframes uparrowmovesenemy {
-  from {
-    top: 11%;
-  }
-
-  to {
-    top: 10.6%;
-  }
-}
-
-@keyframes downarrowmovesenemy {
-  from {
-    top: 38%;
-  }
-
-  to {
-    top: 38.4%;
-  }
 }
